add alert, adjustment, testing bug

+++ public/script/dashboard.js
@@ -25,17 +25,39 @@ function redirectToAddPage() {
 }
 
 function deleteProduct(petID) {
-  fetch(`/pets/delete/${petID}`, {
-    method: "DELETE",
-  })
-    .then((response) => {
-      if (response.ok) {
-        window.location.reload();
-      } else {
-        throw new Error("Failed to delete product");
-      }
-    })
-    .catch((error) => {
-      alert("Failed to delete product. Please try again.");
-    });
+  Swal.fire({
+    title: "Are you sure?",
+    text: "Once deleted, you will not be able to recover this pet!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#d33",
+    cancelButtonColor: "#19a7ce",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      fetch(`/pets/delete/${petID}`, {
+        method: "DELETE",
+      })
+        .then((response) => {
+          if (response.ok) {
+            Swal.fire(
+              "Deleted!",
+              "Your pet has been deleted.",
+              "success"
+            ).then(() => {
+              window.location.reload();
+            });
+          } else {
+            throw new Error("Failed to delete product");
+          }
+        })
+        .catch((error) => {
+          Swal.fire(
+            "Error!",
+            "Failed to delete product. Please try again.",
+            "error"
+          );
+        });
+    }
+  });
 }
